Moved the Compares to Logic Package
