+++ src/PizzaBar/logic/CompareByTime.java
@@ -1,4 +1,4 @@
-package PizzaBar;
+package PizzaBar.logic;
 
 import PizzaBar.products.Order;
 import java.util.Comparator;
